Move roster DJ editing controls from inline table row to modal
The AccountEntry table row was cluttered with interactive controls (role dropdown, capability toggles, inline email editor, password reset, delete buttons). This extracts all editing into an AccountEditModal that opens from a single settings button, making the roster table scannable while keeping all admin functionality accessible.
